+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Under development
 
+- `Element`
+    - `Cookie`
+        - `.decode` -- Handle broken encodings.
+    - `Form`
+        - `.decode` -- Handle broken encodings.
+
 ## 1.2.1 _(July 25, 2015)_
 
 - HTTP